python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
numpy==1.26.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
